Move support file inline with test which uses it.
