+++ test/lib/vedeu/presentation/colour_test.rb
@@ -4,6 +4,32 @@ require 'test_helper'
 
 module Vedeu
 
+  class ParentPresentationColourTestClass
+
+    include Vedeu::Presentation
+
+    attr_reader :colour
+
+    def initialize
+      @colour = { background: '#330000', foreground: '#00aadd' }
+      @parent = nil
+    end
+
+  end # ParentPresentationColourTestClass
+
+  class PresentationColourTestClass
+
+    include Vedeu::Presentation
+
+    attr_reader :parent
+
+    def initialize(attributes = {})
+      @colour = attributes[:colour]
+      @parent = attributes[:parent]
+    end
+
+  end # PresentationTestClass
+
   module Presentation
 
     describe Colour do
