+++ test/test_helper.rb
@@ -96,7 +96,6 @@ end # MiniTest
 require 'mocha/setup'
 require 'vedeu'
 require 'support/helpers/model_test_class'
-require 'support/helpers/presentation_colour_test_class'
 
 if ENV['PERFORMANCE'].to_i == 1
   require 'minitest/reporters'
